+++ src/Resources/app/administration/src/main.js
@@ -1,1 +1,3 @@
 import './module/sw-customer';
+import './component/codecom-freshdesk-sync-reset-button';
+